move chains.js
